Clean up or remaining controllers

+++ app/Actions/Install/ApplyMigration.php
@@ -3,6 +3,8 @@
 namespace App\Actions\Install;
 
 use App\Exceptions\InstallationException;
+use App\Exceptions\Internal\FrameworkException;
+use Illuminate\Contracts\Container\BindingResolutionException;
 use Illuminate\Support\Facades\Artisan;
 
 class ApplyMigration
@@ -10,12 +12,12 @@ class ApplyMigration
 	/**
 	 * Arrayify a string and append it to $output.
 	 *
-	 * @param $string
-	 * @param array $output
+	 * @param string   $string message text which each message separated by newline
+	 * @param string[] $output list of messages
 	 *
-	 * @return array
+	 * @return void
 	 */
-	private function str_to_array($string, array &$output)
+	private function str_to_array(string $string, array &$output): void
 	{
 		$a = explode("\n", $string);
 		foreach ($a as $aa) {
@@ -23,66 +25,65 @@ class ApplyMigration
 				$output[] = $aa;
 			}
 		}
-
-		return $output;
 	}
 
 	/**
-	 * @return bool
+	 * @param string[] $output list of messages
+	 *
+	 * @return void
+	 *
+	 * @throws InstallationException
 	 */
-	public function migrate(array &$output)
+	public function migrate(array &$output): void
 	{
 		Artisan::call('view:clear');
 		Artisan::call('migrate', ['--force' => true]);
 		$this->str_to_array(Artisan::output(), $output);
 
 		/*
-		 * We also double check there is no "QueryException" in the output (just to be sure).
+		 * We check there is no "QueryException" in the output (just to be sure).
 		 */
 		foreach ($output as $line) {
 			if (str_contains($line, 'QueryException')) {
-				// @codeCoverageIgnoreStart
-				return true;
-				// @codeCoverageIgnoreEnd
+				throw new InstallationException('DB migration failed: ' . $line);
 			}
 		}
-
-		return false;
 	}
 
 	/**
-	 * @return bool
+	 * @param string[] $output list of messages
+	 *
+	 * @return void
 	 *
 	 * @throws InstallationException
+	 * @throws FrameworkException
 	 */
-	public function keyGenerate(array &$output)
+	public function keyGenerate(array &$output): void
 	{
 		try {
 			Artisan::call('key:generate', ['--force' => true]);
 			$this->str_to_array(Artisan::output(), $output);
-			// @codeCoverageIgnoreStart
-		} catch (\Exception $e) {
-			$output[] = $e->getMessage();
-			$output[] = 'We could not generate the encryption key.';
-
-			return true;
-		}
-		// @codeCoverageIgnoreEnd
-
-		// key is generated, we can safely remove that file (in theory)
-		$filename = base_path('.NO_SECURE_KEY');
-		if (file_exists($filename)) {
-			if (is_file($filename)) {
-				try {
-					unlink($filename);
-				} catch (\Throwable $e) {
-					throw new InstallationException('Could not remove ' . $filename, $e);
-				}
-			} else {
-				throw new InstallationException('A filesystem object . ' . $filename . ' exists, but is not an ordinary file.');
+			if (!str_contains(end($output), 'Application key set successfully')) {
+				$output[] = 'We could not generate the encryption key.';
+				throw new InstallationException('Could not generate encryption key');
 			}
-		}
 
-		return false;
+			// key is generated, we can safely remove that file (in theory)
+			$filename = base_path('.NO_SECURE_KEY');
+			if (file_exists($filename)) {
+				if (is_file($filename)) {
+					unlink($filename);
+				} else {
+					throw new InstallationException('A filesystem object . ' . $filename . ' exists, but is not an ordinary file.');
+				}
+			}
+		} catch (\ErrorException $e) {
+			// possibly thrown by `unlink`
+			$output[] = $e->getMessage();
+			$output[] = 'Could not remove file `.NO_SECURE_KEY`.';
+			throw new InstallationException('Could not remove file `.NO_SECURE_KEY`', $e);
+		} catch (BindingResolutionException $e) {
+			throw new FrameworkException('Laravel\'s container component', $e);
+		}
 	}
 }
